shill: vpn: Migrate IPAddress for WireGuard services
The local IPv4 overlay address for a WireGuard service used to be stored
in the StaticIPConfig section. After IPv6 is supported for WireGuard,
the local address becomes a list and thus we move it to a specific
property in WireGuard.

This patch add the code to migrate this property before saving and after
loading the service, so that the wg service created either 1) by a
previous version of shill or 2) by Chrome which is still using the
previous property can still work.

BUG=b:266778926
BUG=b:262662603
TEST=FEATURES="test" emerge-$BOARD shill
TEST=`tast run $DUT network.VPNUI.wireguard` with
  https://crrev.com/c/4222730; also checked the WireGuard.IPAddress
  property on the service in the test.

Change-Id: I9fdf92781ac83db942b63ae633ba8d8666b621e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4227892
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>