shill: Move SubscribeToEvents calls into wifi_provider
The NetlinkManager function SubscribeToEvents essentially acts as an
allowlist for certain message types. If a message type has not been
explicitly added via a call to SubscribeToEvents, then those messages
will be ignored at the NetlinkSocket layer. This prevents message
handlers (e.x. those added with AddBroadcastHandler) from receiving
these messages.

This change moves the invocations of SubscribeToEvents from wifi.cc into
wifi_provider.cc. This ensures that we are subscribed to nl80211
mulitcast events even if wifi is currently not enabled in userspace.
This is necessary because with the advent of SoftAP, we now have use
cases for wifi config information (e.x. WiFiPhy capabilities) even when
WiFi is disabled.

It's important to note that calling SubscribeToEvents doesn't actually
mean that Shill will do anything with those events. They still need to
be handled by a message handler. Thus, it should be safe to subscribe to
these events more broadly. I did some work to validate this assertion,
lised in the $TEST section. Another hint that this should be OK is that
the event subscriptions weren't even tied to the lifecycle of the WiFi
device previously. The event subscriptions live in the NetlinkManager
object, and thus are persisted even if the WiFi object is stopped or
destroyed.

This change also makes a small change to a comment to better reflect the
new flow.

BUG=b:266832532
TEST=Add logging statements to log when we receive broadcast messages.
We correctly receive a NEW_WIPHY message on startup, even if WiFi is
disabled. There is no change in the broadcast messages that the WiFi
object receives during boot up and basic WiFi usage.
TEST=FEATURES=test emerge-hatch shill

Change-Id: I97e2a072296cd7a62837dfc0635e0b41a208a548
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4228353
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>