secanomalyd: adds Landlock state to system context
Adds C definitions needed for querying the state of Landlock support
from the kernel. Adds an attribute to the SystemContext class to track
the state of Landlock support and the corresponding method for updating
the attribute using the aforementioned C definitions. Updates the BUILD
file to declare Landlock definitions as a secanomalyd dependency.

BUG=b:268245727
TEST=When the kernel supports Landlock and it is correctly configured,
the attribute reflects that and the version is shown correctly.
TEST=When the kernel does not support Landlock, or when it does but
CONFIG_SECURITY_LANDLOCK=n, we get LandlockState::kNotSupported.
TEST=When the kernel supports Landlock and CONFIG_SECURITY_LANDLOCK=y,
but CONFIG_LSM does not include Landlock i.e. Landlock is not loaded, we
get LandlockState::kDisabled.

Change-Id: I6965f224b1e6e858e5da1ecb08fbf29eb55cb63c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4227520
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Ben Scarlato <akhna@google.com>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>