camera: Use minigbm_create_default_device helper to create gbm device
The helper was added in crrev.com/c/4163418. It first looks for a render
node, if none can be found it then looks for a primary node. To test
cros-camera on display-only drivers like vkms it should be able to use
the primary node.

BUG=b:239110721
TEST=Open camera app on VM with vkms as display

Change-Id: I1dccd13e889cb2e900b6f09cf6caf37645a3eae0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4230936
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Yi Xie <yixie@chromium.org>
Tested-by: Yi Xie <yixie@chromium.org>