dlp: Delay adding files to DB in case it's not ready yet.
Originally DB was initialized synchronously and it was assumed
that no AddFile() request will come before it. With changing
DB initialization to asynchronous, it became possible, but still
unlikely to happen in reality.
However, this sometimes happens in Tast tests that add a file
automatically very quickly after the user session is started.
This leads to the file not being "protected".

BUG=b:265636292
TEST=cros_workon_make --board=<BOARD> --test chromeos-base/dlp

Change-Id: I501b7777940534bd11f478283c70dc77b63f756d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4239437
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>