login: Drop kGuestUserName references
The string based versions of these cryptohome functions are being phased
out in favor of more strongly typed username parameters. This change
gets rid of login manager references to the string constant and replaces
them with GetGuestUsername calls that return the strongly typed value.

When touching these files, I also changed other calls to the libbrillo
functions to also use the strongly typed ones, as we'll be removing them
later too, not just the constant. In most cases this was basically just
adding explicit string->Username conversions, or in some cases the
reverse.

It might actually be a good idea to convert the SessionManager API over
to using the strong types, which would also eliminate most of the manual
conversions here, but that's a much more invasive change than I want to
make to code I'm not specifically familiar with.

BUG=b:263367263
TEST=FEATURES=test emerge-$BOARD chromeos-login

Change-Id: Ic171724c7fd11564c8975fa49766567d05aba7fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4242416
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>