ml_core: install prebuilt opencl cache, update with opencl_cacher
This CL introduces a prebuilt OpenCL cache. It gets installed by the
opencl_cacher tool, which will then run and overwrite the prebuilt files
with the ones generated on the device.

This will prevent the ~15s cold start problem if a user tries to use the
effects before the opencl_cacher has had an opportunity to complete.

BUG=b:257160623
TEST=emerge, deploy and test on brya, nissa and guybrush

Cq-Depend: chrome-internal:5469249
Change-Id: Ib1147a7a5cfe1d83ebf9df9333d818fad8bc327c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4262887
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>