libchromeos-ui: Explicitly specify Floss availability
Before, the Floss flag would be used to both describe whether or not
the user had enabled Floss, and whether or not Floss was available on
the current machine. This caused a bug where a user disabling Floss via
UI would also mark Floss as unavailable on the current machine.

This change fixes the issue by separating the latter functionality into
a new flag.

BUG=b:268585479
TEST=manual

Change-Id: I1e2dc115d4e1046455df4cefaec545163040d8a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4246777
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Sarvesh Kalwit <sarveshkalwit@google.com>
Commit-Queue: Sarvesh Kalwit <sarveshkalwit@google.com>