login: Clear out other string username functions uses
As a followup to chromium:4242416 this change also gets rid of the other
uses of libbrillo username functions that are string-based, adding in
the appropriate type conversions in order to get them to use the
strongly typed overloads.

To verify that there were no more remaining references I temporarily
applied a patch that deleted the libbrillo functions and verified that
everything still built & passed.

BUG=b:263367263
TEST=FEATURES=test emerge-$BOARD chromeos-login

Change-Id: Ia7ed84936cb344d4c4ce04d2e0a7a77a3f534537
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4246998
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>