libbrillo: Remove deprecated string overloads
This removes the string overloads of the various username functions,
just leaving behind the versions that take StrongAlias types.

One exception here is IsSanitizedUserName which I realized after trying
to make use of it, doesn't really make sense to accept a strong type
since you don't actually know if the input is a username or not. So I
just reverted the new version instead.

BUG=b:263367263
TEST=FEATURES=test emerge-${BOARD} libbrillo
TEST=USE=fuzzer emerge-${BOARD} libbrillo
TEST=tested attestation
TEST=tested crash reporter
TEST=tested login manager
TEST=tested cryptohome
TEST=tested arc-data-snapshotd
TEST=tested arcvm-data-migrator
TEST=tested arc-setup

Change-Id: Ica0a8d669c9adc435565a8412d4747023d46be5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4250959
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>