arc: Create span using unsigned length for libvda
A recent CL https://crrev.com/c/4219636 checks that span sizes are
unsigned.

This CL changes the type of the length of a span in arc/libvda to be
unsigned to meet the requirements. It is required to libchrome uprev
r1103195.

BUG=None
TEST=CQ

Change-Id: I6bf6126d424990b7bcefef0eb78c2461d5360cc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4246533
Reviewed-by: Yury Khmel <khmel@google.com>
Auto-Submit: Cici Ruan <cuicuiruan@google.com>
Commit-Queue: Cici Ruan <cuicuiruan@google.com>
Tested-by: Cici Ruan <cuicuiruan@google.com>