libbrillo: Disable enterprise rollback on reven
Crossystem, automatic re-enrollment and pstore do not work on reven as
expected. Disable enterprise rollback by default if requested in a
Flex device.

We may allow enabling the feature at runtime in the future but the USE
flag will still be needed to fix/bypass the features that do not work
as expected.

BUG=b:262689487
TEST=1) FEATURES=test emerge-$BOARD libbrillo for hatch and reven.
        Tests passes in both cases.
     2) Emerge libbrillo for hatch/jacuzzi and reven. Verify rollback
        is enabled for hatch/jacuzzi and disabled for reven. Log shows
        expected message.

Cq-Depend: 4247361, chrome-internal:5461530
Change-Id: If9de533267702fe9bf40a65b94059a211c0abb2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4249319
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>