cryptohome: Rewrite KeysetManagement test to use EncryptEx
This CL is essentially a rewrite of some tests. A couple tests were
deleted as well. They are:
1. ReSaveOnLoadNoReSave was deleted but is tested extensively in
   KeysetManagement::ReSaveOnLoadTestRegularCreds
2. AreCredentialsValid was deleted but is tested extensively in
   GetValidKeysetNonExistentLabel and GetValidKeysetInvalidCreds in
   KeysetManagement unit tests.

Also, in AddKeysetWithKeyBlobsSaveFail the encrypt step was deleted
because it did not add any value to the test and the step was not
essential to what is being tested, which is save operation.

Rest of the changes should be very cosmetic. This CL does not intend on
deleting GetValidKeyset references, even though some are now deleted
through this test.

BUG=b:229825202
BUG=b:261022410
TEST=FEATURES=test emerge-zork cryptohome

Change-Id: I0947640fbe9f0801430c9d1930fceb7f0209e1d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4249182
Reviewed-by: John Admanski <jadmanski@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>