libchromeos-ui: Set SW cursors flag on herobrine
On certain QC SoCs, the display controller uses multiple planes for
sufficiently high resolution displays. Planes are limited, so in
some scenarios (e.g. attaching a 4k external monitor), we run out of
planes to have a dedicated hardware cursor and instead display no
cursor at all.

This CL instructs Chrome to use software cursors on herobrine. We
use the base board name directly as opposed to say, querying for the
SoC name because there can be multiple variants of a SoC in the
different SKUs of a board (e.g. trogdor had both SC7180 and SC7180P).
This simplifies the logic a bit and doesn't require us to update this
file every time a new herobrine variant is added.

BUG=b:243597541
TEST=Flash to device and verify command line args in chrome://gpu

Change-Id: I6c11566de5897d6d7160ffcbec7b99762a5197da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4246861
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Brian Ho <hob@chromium.org>
Commit-Queue: Brian Ho <hob@chromium.org>