shill: cellular: expect APN type in CellularCapability3gpp::Connect()
The Connect() method in CellularCapability3gpp is updated so that it
can be called independently for different APN types: either DEFAULT or
DUN.

The logic is now based on using a new struct that holds the
information of a given connection attempt: the user provided
ResultOnceCallback and the APN list being processed. These structs are
stored in a map per APN type, and it is ensured there is only max one
attempt per APN type.

Additional connection attempts happening while one is already ongoing
are now forbidden, to ensure the APN try list of the ongoing attempt
is not updated unexpectedly.

The logic storing the last good APN is exclusively applied to the case
of the DEFAULT APN type, and ignored for DUN.

BUG=b:269342469
TEST=CQ passes

Change-Id: Id44679b08537d8d9251e1e7bb92ad04c69267175
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4258485
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>