camera: Create span using unsigned length
A recent CL https://crrev.com/c/4219636 checks that span sizes are
unsigned.

This CL changes the type of the length of a span to be unsigned to meet
the requirements. It is required for libchrome uprev r1103195.

BUG=None
TEST=emerge-${BOARD} cros-camera-libs libchrome of new and old.
TEST=CQ

Change-Id: Ic8193408001ff45c96a82006d86e55c9ff72d584
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4252160
Tested-by: Cici Ruan <cuicuiruan@google.com>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Auto-Submit: Cici Ruan <cuicuiruan@google.com>
Commit-Queue: Cici Ruan <cuicuiruan@google.com>