runtime_probe: Log the output of PostHelperEvalImpl()
We used to log the output of InvokeHelper(), which doesn't contain the
output of PostHelperEvalImpl() and thus may be different from the full
result of the probe function. To make the logging clear for debugging,
modify |PrivilegedProbeFunction| to log the result after
PostHelperEvalImpl() is called.

BUG=b:269208280
TEST=FEATURES=test emerge-${BOARD} runtime_probe
TEST=Manually test on DUT

Change-Id: Ib7ba7831f25c85ea20c065256b94ebeae26aa54f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4270870
Reviewed-by: Kevin Lin <kevinptt@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>