shill: cellular: expect APN type in CellularCapability3gpp::GetActiveBearer()
The logic processing active bearer objects is updated to consider the
case of having additional active bearers of different APN types. The
limit of max one bearer per APN type is still imposed; if more than
one found for the same type they will be ignored (instead of crashing
shill as it was done before).

This logic also limits the DBus properties changes monitoring only
to the Bearer object of Default APN type, as we only use it to be
reported of uplink/downlink speed changes.

BUG=b:269342469
TEST=CQ passes

Change-Id: I072c2a14973fe8cfb7ac14810e77d34d8860b997
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4258486
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>