cryptohome: Disable partial backup VK creation
If a session is started for a user that has a mix of factors with and
without backup VKs, disable creation of any new backup VKs.

This is relevant in situations where all your factor types nominally
support backups but you are missing some, for example because you added
and removed a recovery factor. It also has some side effects if a backup
VK gets corrupted but those might actually be good (e.g. it will disable
backups instead of just failing).

BUG=b:262632342
TEST=FEATURES=test emerge-${BOARD} cryptohome

Change-Id: I6ee1718e09974e08cea63d8c161a451fe2d17f83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4257285
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>