power: Don't disable battery sustainer unless it's set
This prevents some log spam in powerd due to the battery sustainer
repeatedly being stopped despite the lack of support of some platforms.
This adds a check to only disable it if it was set to a non-disabled
value (which won't happen if the feature isn't supported).

BUG=None
TEST=unittests

Change-Id: Ic35680508ccddb6677b48f12aa9e5acb6bd9e71f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4261699
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>