cryptohome: Make the on-timeout callback settable
This probably goes against my usual design suggestions for objects,
where I'm usually suggesting moving such state into the constructor, but
this callback is a bit different.

The problem is that an on-timeout callback is used to manage the
lifetime of the AuthSession, but making it a constructor argument means
that you now can't easily transfer ownership of the objects because if
the owner wants to use the callback for deletion they also need to be
the creator of it. This makes it difficult to for example transfer
ownership of AuthSession objects to the AuthSessionManager.

BUG=b:263259501
TEST=FEATURES=test emerge-${BOARD} cryptohome

Change-Id: Ib8e65f368503a238b9616d6891da7b071674e4b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4257594
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>