shill: wifi: HandleEnsuredScan on idle transition
HandleEnsuredScan is meant to be called on all phystate idle
transitions. This change makes sure that even when the transition is
idle -> idle (nothing changes), the HandleEnsuredScan method will still
be called.

Resolves potential looping behavior when the device becomes disabled or
supplicant becomes unavailable between ensuring the scan and the scan
being handled.  In this case, ConnectToBestServices is called in the
hopes that another device will be able to connect.

BUG=b:242940677
TEST=FEATURES="test" USE="-cros-debug" emerge-$BOARD shill

Change-Id: Idc5a02435ddafe6cf38f37704ad3ccf2466d8360
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4258115
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: David Ruth <druth@chromium.org>
Tested-by: David Ruth <druth@chromium.org>