shill: cellular: skip link establishment in modem connected transition test
The ModemStateChangeValidConnected test was working by luck. This test
was making the device state transition to Connected successfully, but
it would then fail to go on because no active Bearer was detected
during the link establishment.

This failure triggered a Disconnect() but then it was a no-op mostly
because there was no Simple interface proxy to run the disconnect, and
the ResultCallback would be ignored.

Fix the test by completely avoiding the link establishment procedure,
which has its own tests.

BUG=None
TEST=cros_workon_make --board=${BOARD} --test shill

Change-Id: I3e3771fcc6215f1b4fec93aec88da5f3522a3713
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4258489
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>