cryptohome: Remove all backup keysets when recovery is added
Current AuthSession code removes the backup keyset for a factor if the cryptohome recovery is added for the user account AND the factor is being updated. In this case user may have other factors that have backup keysets, leaving an asymmetric factor backing store. This is not problematic case and handled in the code, but due the recent recovery bugs caused by the uss migration and backup keysets it is safer to create less number of asymmetric cases between different factors. This CLs removes all the backup keysets when a cryptohome recovery factor is added making all the factors backed only by USS.

BUG=b:262632342
TEST=Unittests pass.

Change-Id: Ie7d2ebc782958f070b262a518e30d50a1df46910
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4256607
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Betul Soysal <betuls@google.com>