diagnostics: Report diagnostic result kCancelling as kCancelled in UMA
The status of |SubprocRoutine| won't be updated if no client queries for
it. Usually the clients stop querying the routine status after
cancelling it. This results in the routine stuck in the cancelling
status so the result is never reporting to UMA. A workaround is to treat
the cancelling status as cancelled when reporting it to the UMA. This
should be fine since the routine should eventually be cancelled.

BUG=b:269409037, b:243335991
TEST=USE="-cros-debug" FEATURES=test emerge-volteer diagnostics
TEST= 1. Deploy diagnostics to a DUT
      2. Open the Diagnostics App, start CPU tests and stop it.
      3. Verify the record for diagnostics in chrome://histograms.

Change-Id: Ia6ae0fee5fe920280979379fda85710489942ff5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4258606
Reviewed-by: Byron Lee <byronlee@chromium.org>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>