rmad: Fix |TransitionNextState_MissingHandler| test
We currently expect to stay in the original state when the state handler
for the next state is missing. However, the corresponding tests passed
accidentally when we made this change (not making |rmad| crash).

This CL is to fix the related tests to make correct expectations.

BUG=b:269383407
TEST=FEATURES=test emerge-octopus rmad

Change-Id: Ia00c828a05e66402e83f760c315a3df4beb9310e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4259251
Auto-Submit: Jeff Lin <jeffulin@google.com>
Tested-by: Jeff Lin <jeffulin@google.com>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>