libchromeos-rs: Pass BUFFER_SIZE to C with .into()
BUFFER_SIZE is declared as usize, then passed to a C function that
expects a size_t. Depending on which bindgen version is used, this
becomes either a c_ulong or a usize in Rust. The existing code
uses "as c_ulong", but this does not work if bindgen gives us a
usize, because c_ulong is not guaranteed to fit in a usize.
Fortunately, BUFFER_SIZE is a constant that isn't large enough
to run into this problem. This change declares BUFFER_SIZE as
u16 and uses .into() to pass it to C, which will convert to
c_ulong or usize as appropriate.

BUG=b:269573224
TEST=emerge with both bindgen-0.59.2 and bindgen-0.63.0

Change-Id: I4fcf88c7b5d4a01640e9ee64a9a1bee48e2c3793
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4260105
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>