platform2: Create span using unsigned length
A recent CL https://crrev.com/c/4219636 checks that span sizes are
unsigned.

This CL changes the type of the length of a span to be unsigned to meet
the requirements. It is required for removing backforward patch in
libchrome uprev r1103195.

BUG=None
TEST=CQ

Change-Id: I0cc4880b0ad206996983046d6c4eba2fd4f8fde2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4263549
Auto-Submit: Cici Ruan <cuicuiruan@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Cici Ruan <cuicuiruan@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Andreea Costinas <acostinas@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Cici Ruan <cuicuiruan@google.com>