shill: net: Create IPAddress from a prefix for a given family
This patch extends the current IPAddress::CreateFromPrefixString()
function to accept a |family| parameter. When |family| is specified,
this function will only try to parse the input string for that family.

The usages of IPAddress::SetAddressAndPrefixFromString() will be
replaced by this function in a follow-up patch. This change makes this
function easier to use in the follow-up refactor.

BUG=b:269983153
TEST=FEATURES="test" emerge-${BOARD} shill-net

Change-Id: I406327ba0102ec1f33884bd3956cad7e5cd216fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4263331
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>