shill: net: Hide IPAddress::SetAddressAndPrefixFromString()
This function is no longer a public interface. The static method
CreateFromPrefixString() should be used instead.

Also remove the error log in the function. The caller should log it on
failure which will have more context. This can also avoid the error
logs when CreateFromPrefixString() is used since it may try to parse the
string as both v4 and v6.

BUG=b:269983153
TEST=FEATURES="test" emerge-${BOARD} shill-net

Change-Id: Ifda0bf8e7e5f552cdea602f59c5886fb014f4c4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4263333
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>