virtual_file_provider: Disable Landlock in minimalistic-mountns profile
CL:4233938 enabled Landlock in minijail's minimalistic-mountns profiles
(previously it was enabled only when there is no -k mount).

As a result, some operations like mount are disabled in such profiles,
causing certain services not to work properly.

virtual-file-provider is one of such services; now arc.Fsp is broken due
to the change, which is fixed by disabling Landlock by adding minijail's
no-fs-restrictions option in virtual-file-provider-jaild.sh.

Note that virtual-file-provider-jailed-arcvm.sh does not need to be
modified, because it does not use a minimalistic-mountns profile.

BUG=b:265721189
BUG=b:269374038
BUG=b:269507578
TEST=arc.Fsp

Change-Id: I3c74b6f1f7efed0383c4bf93af38b52bd199375b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4269713
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Momoko Hattori <momohatt@chromium.org>