camera: use canonical EffectsConfig, move mojom into camera stack
This CL removes the manually managed copy of EffectsConfig from the
ChromeOS source code, and depends on it being installed into /usr/local
by the ml-core-internal DLC. This removes the potential for divergence
between Google3 and ChromeOS for the EffectsPipeline system.

This CL also moves the mojom definitions from ml_core to the camera
stack, which is the only user / consumer of it. It fits better there.

BUG=b:237964122
TEST=build & deploy to brya DUT, test effects still work

Cq-Depend: chrome-internal:5498609
Change-Id: I7dbe371dc9f1da1b1cf0f5534c4187e6b74d44c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4273445
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>