shill: Simplify check_portal_list property
Current implementation of check_portal_list loads existing manager
property when shill is started and is reset to default by Chrome when
restart, which can make some tests fail.

This patch simplifies this property and changed it in the way that
this patch removes the --portal-list command-line flag used by oobe,
but no longer needed because the decouple of portal setting and oobe.
Now when shill starts, the CheckPortalList property will be loaded from
the default profile, or set to the default value if the default profile
does not exist.

In shill patch following changes were made:
Remove support for --portal-list on shill command line.

BUG=b:261820123,b:265806000
TEST=tast run $DUT network.AlwaysOnVPNRelogin.*

Change-Id: I7228cc512ca1ce0b5a4763f14d6ded56deffefe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4270434
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Chuwei He <chuweih@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Chuwei He <chuweih@google.com>