vm_tools: Move vm_* protos so they can be included more easily
Conventionally system_api protos for dbus would be generated into a
"proto_bindings" dir. This prevents us from using the import directive
in proto files, since the resulting .pb.h would not be present (in the
correct place).

We solve this by conforming the proto_in_dir and proto_out_dir to the
same value for all vm_* protos. We align with the "arc" protos on "dbus"
and "include", since we already include some of their definitions in our
protos.

We update the relevant #includes in vm_tools code.

BUG=b:270091670
TEST=emerge-volteer system_api vm_host_tools

Change-Id: Idc61a0c5b3bfa551e6b41bd5395ab9ccd068839d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4274185
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>