lorgnette: Log more long-running operations
There are several scanning steps that can take a long time.  If
something fails in one of these steps, lorgnette sends an error to the
frontend, but doesn't log anything.  This makes it difficult to see what
went wrong when reviewing test logs after the fact.  To make these
easier to debug, start logging some additional info before and after
some of the long stages, as well as whenever an active scan job is
created or deleted.

BUG=b:268216681
TEST=tast run scanapp.Scan and review logs

Change-Id: I162c156111e144fa1e6f622ed38633f8a7b5b1ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4282137
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mason Wilde <masonwilde@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>