vm_tools: Refactor garcon::HostNotifier
HostNotifier has a mix of static and non-static methods. Despite a token
and gRPC stub being bound to a HostNotifier instance, the static methods
cannot use them.

A HostNotifier must now be instantiated to call its methods. All methods
now use the security token and gRPC stub bound to the object. The
shutdown closure is only provided at initialisation of the server,
rather than at instantiation, as it is not needed in client mode.

The HostNotifier instance is now also provided to the D-bus service
objects that use its methods as well.

BUG=b:270485596
TEST=Boot Crostini and Bruschetta, ensure garcon is operational

Change-Id: Id893b33f0ca91472d4997ad1136af8777e327391
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4286240
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: James Ye <jamesye@google.com>
Tested-by: James Ye <jamesye@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>