login: Allow featured to modify /sys/fs/cgroup/cpu/chrome_renderers
For featured to modify:
  /sys/fs/cgroup/cpu/chrome_renderers/cpu.uclamp.min
  /sys/fs/cgroup/cpu/chrome_renderers/foreground/cpu.uclamp.min

root must have write permission. As ui-pre-start does a chown to
chronos, selinux prevents featured (which is root) from modifying it, as
it is prevented by "dac_override" not being set. Instead of giving
"dac_override" permission to featured (which may open up other security
concerns), simply change the group permissions of the cgroups files to
write. This will still allow root to write to these files as well as for
chronos (which was the reason for the chown).

This allows us to do CL:4258120 which updates an experiment with boosts.

BUG=b:248028313
TEST=login and see things update

Change-Id: I1998f4960af7ebe42e4f63eb4666f54d3ae0c465
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4284822
Tested-by: Steven Rostedt <rostedt@google.com>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Auto-Submit: Steven Rostedt <rostedt@google.com>