shill: cellular: improve DisconnectNoProxy test
The ReleaseCapabilityProxies() was basically a no-op because
capability_->proxies_initialized was not set.

The test was passing just because we didn't set a simple proxy with
SetSimpleProxy(), so capability_->simple_proxy was really nullptr
anyway.

Improve the test by first setting the simple proxy, and then calling
ReleaseCapabilityProxies() so that it really deletes the proxy before
calling Disconnect().

BUG=None
TEST=cros_workon_make --board=${BOARD} --test shill

Change-Id: I5c494adfdced25e77ac9f4940330b530a6d6a54d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4286548
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>