debugd: Implement GetFeedbackLogsV3
GetFeedbackLogsV3 is the same as GetFeedbackLogsV2 except that it tries
to run sub task in parallel in order to reduce client wait time and to
reduce the chance of timing out.

A feature flag #enable-get-debugd-log-in-parallel has been added in
a separate CL. When the flag is enabled, GetFeedbackLogsV3 will be used.

The base::PostJob api is better suited for this than the PostTask api,
for the calling thread can also contribute instead of blocking or
running a nested loop.

Further optimization is possible and will be done in followup CLs.

BUG=b:238782324
TEST=manually tested on VM.
  On a Kukui device, time taken reduced to ~9.5 from ~18.5 seconds.

Change-Id: I5fd53406c32381a91c92f761e99268d804b3b73d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4288444
Commit-Queue: Xiangdong Kong <xiangdongkong@google.com>
Tested-by: Xiangdong Kong <xiangdongkong@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>