vm_tools: sommelier: fix gcc -Wstringop-truncation warning on strncpy
Specifically, this warning:

../sommelier.cc:3055:10: warning: ‘char* __builtin_strncpy(char*, const char*, long unsigned int)’ specified bound 108 equals destination size [-Wstringop-truncation]

The message is a little obscure, but is referring to the fact that
the destination string will not be null-terminated if the source
string is longer than the buffer size; see `man strncpy`.

The compiler is smart enough to silence the warning if we manually
null-terminate the string. This will of course truncate the string
at 107 characters if it does overflow, but that seems preferable to
storing an unterminated string.

BUG=none
TEST=compile sommelier with gcc -Wstringop-truncation

Change-Id: I0bfaa19127f44f99b8395ab41dda6fd109d04ed1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4290568
Auto-Submit: Chloe Pelling <cpelling@google.com>
Commit-Queue: Justin Huang <justinhuang@google.com>
Tested-by: Chloe Pelling <cpelling@google.com>
Reviewed-by: Justin Huang <justinhuang@google.com>